mem-ruby: Fix warmup with backing store
When warming up with Ruby backing store enabled, you cannot access both
the backing store and the mem_interface with the packet. This change now
accesses one or the other. The downside to this approach is that with
this change you cannot restore a checkpoint or switch CPUs and disable
backing store.

Change-Id: I2d9f542a759d47fea571abd0d3141ce5996a81c3
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>