python: Allow SimObjectVector to match proxy
There are proxies (like System.memories) which want to find *all* of the
SimObjects that are of that type. However, the code currently cannot
deal with SimObjectVectors. This change now checks for SimObjectVectors
and updates the proxy list with each SimObject in the vector. It's been
tested with System.memories and a VectorParam.SimpleMemory.

This is needed to allow multiple discontinuous phys_mem parameters for
Ruby's backing store.

Change-Id: I937fbfc41a8a611b1b3a491caec41e8a9dcb0947
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>