mem-ruby: Allow multiple backing store ranges
This change enables Ruby to have multiple backing store ranges. The
phys_mem parameter is now a list of simple memory objects, one for each
discontiguous range.

This is needed to have a second memory range above 4GiB for x86
simulation.

Note that this is currently not used in Ruby.py, but it should be
straightforward to set up for anyone's specific use case.

Change-Id: I591a2b82a06bec69f583ba21f7267b6f535f48b1
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>